Returns unique link
